requirements.txt
streamlit==1.31.1
langchain==0.1.16
langchain-community==0.0.38
langchain-groq==0.1.0
youtube-transcript-api==0.6.2
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
python-dotenv==1.0.1
pypdf==4.0.2
markdown==3.5.2
groq==0.4.2
google-generativeai==0.3.2
tiktoken==0.6.0
